@@ -3,22 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
     "axiom": "^0.1.6",
-    "axios": "^0.19.2",
-    "formik": "^2.2.5",
-    "node-sass": "^4.14.1",
-    "npm": "^6.14.9",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-image": "^2.4.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.3.0"
+    "axios": "^1.1.3",
+    "formik": "^2.2.9",
+    "node-sass": "^7.0.3",
+    "npm": "^8.19.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-image": "^4.0.3",
+    "react-router": "^6.4.2",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
-    "start": "SET PORT=4500 && react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
