@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 import MenuItem from '../menu-item/menu-item.component';
 
@@ -15,19 +16,19 @@ class Directory extends React.Component {
                   title: 'Fretboard/Note Trainer',
                  imageUrl: 'https://miro.medium.com/max/3464/1*oJBvXyzRk_LFII8UcGS-1g.png',
                   id: 1,
-                  linkUrl: 'fretboardnotetrainer'
+                  linkPath: 'fretboardnotetrainer'
                 },
                 {
                   title: 'Sight Reading',
                   imageUrl: 'https://image.shutterstock.com/image-vector/music-note-design-element-doodle-260nw-616470641.jpg',
                   id: 2,
-                  linkUrl: 'sightreading'
+                  linkPath: 'sightreading'
                 },
                 {
                   title: 'Learn Theory',
                   imageUrl: 'https://i.stack.imgur.com/XNCwK.png',
                   id: 3,
-                  linkUrl: 'learntheory'
+                  linkPath: 'learntheory'
                 }
               ]   
         }
@@ -40,7 +41,11 @@ class Directory extends React.Component {
                 {
                 
                 this.state.sections.map(({ id, ...sectionProps }) => (
-                    <MenuItem key={id} {...sectionProps}/> ))
+                    // <Link to={sectionProps.linkPath}>
+                      <MenuItem key={id} {...sectionProps}/>
+                    // </Link>
+                    
+                    ))
                 }
             </div>
           
