@@ -1,16 +1,15 @@
 import React from 'react';
-import {withRouter} from 'react-router-dom';
-
 import './menu-item.styles.scss';
+import {Link} from 'react-router-dom';
 
 //This component returns the menu-item to the directory
-const MenuItem = ( { title, imageUrl, size, history, linkUrl, match } ) => (
+const MenuItem = ( { title, imageUrl, size, linkPath, match } ) => (
     // <div className={`menu-item`} onClick={() => history.push(`${match.url}${linkUrl}`)}>
+    <Link to={linkPath}>
         <div  
             className={`menu-item`}
-            
             //took out ${match.url} once Log In page was added. 
-            onClick={() => history.push(`${linkUrl}`)}
+            // onClick={(linkPath) => history.push(`${linkPath}`)}
             style={{
                 backgroundImage: `url(${imageUrl})`
             }}
@@ -18,9 +17,11 @@ const MenuItem = ( { title, imageUrl, size, history, linkUrl, match } ) => (
             <div className='content'>
                 <h1 className='title'>{title}</h1>
             </div>
-    </div>
+        </div>
+    </Link>
+
 
     
 );
 
-export default withRouter(MenuItem);
+export default MenuItem;
